{Etape 4 LB} Petit ajout sur la creation pour etre toujours iso avec la base

+++ src/RestFull/RestFullBundle/Controller/BakCtrl.php
@@ -39,7 +39,7 @@ class UserController extends Controller
         );
     }
 
-     /**
+    /**
      * return a status 200.
      * @Route("/user/", name="home")
      * @Route("/users/", name="homes")
@@ -48,42 +48,29 @@ class UserController extends Controller
      */
     public function homeAction(Request $oRequest)
     {
-    
+	
         $oEm = $this->getDoctrine()->getManager();
-        $oUserRepository = $this->getDoctrine()->getManager()->getRepository("RestFullRestFullBundle:User");
-        $aUserFromJson  = json_decode($oRequest->getContent(), true);
-        
+	var_dump(json_decode($oRequest->getContent(), true));
         if ($oRequest->isMethod('post')) {
-          
-            /* $aUsers = $this->getDoctrine()->getManager()
+            $aUsers = $this->getDoctrine()->getManager()
                 ->getRepository("RestFullRestFullBundle:User")
-                ->findBy(array('id'=>'DESC'));*/
-            
-            //var_dump($aUserFromJson['lastname']);
-            //$iNbUsers = count($aUsers);
-            
-            // Recuperation du dernier di de la base.
-            $aLastUserId = $oUserRepository->getLastUserId();
-            $iLastUserId = (int) $aLastUserId[1];
-            
+                ->findAll();
+
+            $aNbUsers = count($aUsers);
             $oUser = new User();
-            /*$oUser->setId($aNbUsers+1);
-                    $oUser->setlastname($oRequest->request->get('lastname'));
-                    $oUser->setFirstname($oRequest->request->get('lastname'));
-                    $oUser->setEmail($oRequest->request->get('email'));
-                    $oUser->setRole($oRequest->request->get('role'));
-                    $oUser->setPassword($oRequest->request->get('password'));*/
-            $oUser->setId($iLastUserId+1);
-            $oUser ->setLastname($aUserFromJson['lastname']);
-            $oUser ->setFirstname($aUserFromJson['firstname']);
-            $oUser ->setEmail($aUserFromJson['email']);
-            $oUser ->setRole($aUserFromJson['role']);
-            $oUser ->setPassword($aUserFromJson['password']);
-	                
-            $oEm->persist($oUser);
-            $oEm->flush($oUser);
-        }
-	die('created');
+            $oUser->setId($aNbUsers+1);
+            $oUser->setlastname($oRequest->request->get('lastname'));
+            $oUser->setFirstname($oRequest->request->get('lastname'));
+            $oUser->setEmail($oRequest->request->get('email'));
+            $oUser->setRole($oRequest->request->get('role'));
+            $oUser->setPassword($oRequest->request->get('password'));
+
+            $em->persist($oUser);
+            $em->flush();
+	   
+	    $this->getResponse()->setStatusCode('024');
+        };
+
         $response = new Response('Content', 200, array('content-type' => 'text/html'));
         return $response;
     }
@@ -276,108 +263,41 @@ class UserController extends Controller
 
         return $form;
     }
-
-
-
-
-
-     /**
+    /**
      * Edits an existing User entity.
      *
-     * @Route("/user/{id}/", name="user_update")
-     * @Route("/users/{id}/", name="user_update")
+     * @Route("/user/{id}", name="user_update")
+     * @Route("/users/{id}", name="user_update")
      * @Method("PUT")
      * @Template("RestFullRestFullBundle:User:edit.html.twig")
      */
     public function updateAction(Request $request, $id)
     {
         $em = $this->getDoctrine()->getManager();
-        $oUserRepository = $this->getDoctrine()->getManager()->getRepository("RestFullRestFullBundle:User");
 
-        $aUserFromJson  = json_decode($request->getContent(), true);
-        //var_dump("<PRE>",$aUserFromJson);die;
-        
-        // verification et bind des donnees utilisateur du flux json avec celui de la bdd.
-        //var_dump($aUserFromJson);die;
-        
-        //$oUserRepository->getUserByAttrib($aUserFromJson);
-        
         $entity = $em->getRepository('RestFullRestFullBundle:User')->find($id);
-        
-        
+
         if (!$entity) {
             throw $this->createNotFoundException('Unable to find User entity.');
         }
-        
-        $this->updateUserAboutJsonStream($entity, $aUserFromJson);
-        /*
-                $deleteForm = $this->createDeleteForm($id);
-                $editForm = $this->createEditForm($entity);
-                $editForm->handleRequest($request);
 
-                if ($editForm->isValid()) {
+        $deleteForm = $this->createDeleteForm($id);
+        $editForm = $this->createEditForm($entity);
+        $editForm->handleRequest($request);
 
-                    $em->flush();
-                    return $this->redirect($this->generateUrl('home', array('id' => $id)));
-                }
+        if ($editForm->isValid()) {
 
-                return array(
-                    'entity'      => $entity,
-                    'edit_form'   => $editForm->createView(),
-                    'delete_form' => $deleteForm->createView(),
-                );*/
-        
-    }
-    
-    private function updateUserAboutJsonStream($entity, $aUserFromJson) 
-    {
-        //var_dump("<PRE>",$entity, $aUserFromJson);die;
-       // var_dump($entity);
-        /*foreach ($aUserFromJson as $sIndexAttrib => $sValueAttrib)
-        {
-            $attrib = ucfirst( $sIndexAttrib );
-            
-            $methode = "set".$attrib."('$sValueAttrib')";
-            //var_dump($methode);die;
-            var_dump($entity->$methode);die;
-            //$entity->set."$attrib".($sValueAttrib);
-        }*/
-        foreach ($aUserFromJson as $sIndexAttrib => $sValueAttrib)
-        {
-            if ("firstname" == $sIndexAttrib) 
-            {
-                //die('firstname ok');
-                $entity->setFirstName($sValueAttrib);
-            }
-            if ("lastname" == $sIndexAttrib) 
-            {
-                //die('lastname ok');
-                $entity->setLastName($sValueAttrib);
-            }
-            if ("email" == $sIndexAttrib) 
-            {
-                //die('lastname ok');
-                $entity->setEmail($sValueAttrib);
-            }
-            if ("role" == $sIndexAttrib) 
-            {
-                //die('role ok');
-                $entity->setRole($sValueAttrib);
-            }
-            if ("password" == $sIndexAttrib) 
-            {
-                //die('password ok');
-                $entity->setPassword($sValueAttrib);
-            }
+            $em->flush();
+            return $this->redirect($this->generateUrl('home', array('id' => $id)));
         }
-        
-        $oEm = $this->getDoctrine()->getManager();
-        
-        $oEm->persist($entity);
-        $oEm->flush();die;        
-    }
 
-     /**
+        return array(
+            'entity'      => $entity,
+            'edit_form'   => $editForm->createView(),
+            'delete_form' => $deleteForm->createView(),
+        );
+    }
+    /**
      * Deletes a User entity.
      *
      * @Route("/user/{id}", name="user_delete")
@@ -386,12 +306,38 @@ class UserController extends Controller
      */
     public function deleteAction(Request $request, $id)
     {
-	
-        $em = $this->getDoctrine()->getManager();
-        $entity = $em->getRepository('RestFullRestFullBundle:User')->find($id);
+        $form = $this->createDeleteForm($id);
+        $form->handleRequest($request);
 
-        $em->remove($entity);
-        $em->flush();
-        die('done');
+        if ($form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $entity = $em->getRepository('RestFullRestFullBundle:User')->find($id);
+
+            if (!$entity) {
+                throw $this->createNotFoundException('Unable to find User entity.');
+            }
+
+            $em->remove($entity);
+            $em->flush();
+        }
+
+        return $this->redirect($this->generateUrl('user'));
+    }
+
+    /**
+     * Creates a form to delete a User entity by id.
+     *
+     * @param mixed $id The entity id
+     *
+     * @return \Symfony\Component\Form\Form The form
+     */
+    private function createDeleteForm($id)
+    {
+        return $this->createFormBuilder()
+            ->setAction($this->generateUrl('user_delete', array('id' => $id)))
+            ->setMethod('DELETE')
+            ->add('submit', 'submit', array('label' => 'Delete'))
+            ->getForm()
+        ;
     }
 }
